+++ packages/chakra/jest.config.mjs
@@ -3,5 +3,6 @@ import common from '../../jest.common.mjs';
 /** @type {import('@jest/types').Config.InitialOptions} */
 export default {
   ...common,
+  coveragePathIgnorePatterns: [...common.coveragePathIgnorePatterns, 'src/snippets'],
   displayName: 'chakra',
 };