+++ packages/chakra/src/index.tsx
@@ -1,12 +1,13 @@
-import {
-  ChevronDownIcon,
-  ChevronUpIcon,
-  CloseIcon,
-  CopyIcon,
-  LockIcon,
-  UnlockIcon,
-} from '@chakra-ui/icons';
 import * as React from 'react';
+import {
+  FaChevronDown,
+  FaChevronUp,
+  FaCopy,
+  FaGripVertical,
+  FaLock,
+  FaLockOpen,
+  FaTimes,
+} from 'react-icons/fa';
 import type {
   ControlElementsProp,
   FullField,
@@ -15,13 +16,11 @@ import type {
 } from 'react-querybuilder';
 import { getCompatContextProvider } from 'react-querybuilder';
 import { ChakraActionElement } from './ChakraActionElement';
-import { ChakraDragHandle } from './ChakraDragHandle';
 import { ChakraNotToggle } from './ChakraNotToggle';
 import { ChakraValueEditor } from './ChakraValueEditor';
 import { ChakraValueSelector } from './ChakraValueSelector';
 
 export * from './ChakraActionElement';
-export * from './ChakraDragHandle';
 export * from './ChakraNotToggle';
 export * from './ChakraValueEditor';
 export * from './ChakraValueSelector';
@@ -29,22 +28,22 @@ export * from './ChakraValueSelector';
 export const chakraControlElements: ControlElementsProp<FullField, string> = {
   actionElement: ChakraActionElement,
   valueSelector: ChakraValueSelector,
-  dragHandle: ChakraDragHandle,
   notToggle: ChakraNotToggle,
   valueEditor: ChakraValueEditor,
 };
 
 export const chakraTranslations: Partial<Translations> = {
-  removeGroup: { label: <CloseIcon /> },
-  removeRule: { label: <CloseIcon /> },
-  cloneRuleGroup: { label: <CopyIcon /> },
-  cloneRule: { label: <CopyIcon /> },
-  lockGroup: { label: <UnlockIcon /> },
-  lockRule: { label: <UnlockIcon /> },
-  lockGroupDisabled: { label: <LockIcon /> },
-  lockRuleDisabled: { label: <LockIcon /> },
-  shiftActionDown: { label: <ChevronDownIcon /> },
-  shiftActionUp: { label: <ChevronUpIcon /> },
+  dragHandle: { label: <FaGripVertical /> },
+  removeGroup: { label: <FaTimes /> },
+  removeRule: { label: <FaTimes /> },
+  cloneRuleGroup: { label: <FaCopy /> },
+  cloneRule: { label: <FaCopy /> },
+  lockGroup: { label: <FaLockOpen /> },
+  lockRule: { label: <FaLockOpen /> },
+  lockGroupDisabled: { label: <FaLock /> },
+  lockRuleDisabled: { label: <FaLock /> },
+  shiftActionDown: { label: <FaChevronDown /> },
+  shiftActionUp: { label: <FaChevronUp /> },
 };
 
 export const QueryBuilderChakra: QueryBuilderContextProvider = getCompatContextProvider({