+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `@react-querybuilder/chakra` now supports Chakra UI version 3 and no longer supports version 2. For Chakra UI version 2 support, use `@react-querybuilder/chakra2`.
+- `ChakraDragHandle` has been removed. In Chakra UI version 3 environments, the default `DragHandle` component works without augmentation.
+
 ### Added
 
 - Debug build which registers the internal Redux store with Redux DevTools (more "debug" functionality may be added in the future). To use the debug build, update all `"react-querybuilder"` imports to `"react-querybuilder/debug"`.